+++ railties/lib/rails/generators/app_base.rb
@@ -38,6 +38,10 @@ module Rails
         class_option :skip_keeps,         type: :boolean, default: false,
                                           desc: 'Skip source control .keep files'
 
+        class_option :skip_action_mailer, type: :boolean, aliases: "-M",
+                                          default: false,
+                                          desc: "Skip Action Mailer files"
+
         class_option :skip_active_record, type: :boolean, aliases: '-O', default: false,
                                           desc: 'Skip Active Record files'
 
@@ -164,7 +168,7 @@ module Rails
       end
 
       def include_all_railties?
-        !options[:skip_active_record] && !options[:skip_test_unit] && !options[:skip_sprockets]
+        options.values_at(:skip_active_record, :skip_action_mailer, :skip_test_unit, :skip_sprockets).none?
       end
 
       def comment_if(value)