+++ railties/test/generators/plugin_generator_test.rb
@@ -140,6 +140,20 @@ class PluginGeneratorTest < Rails::Generators::TestCase
     end
   end
 
+  def test_app_generator_without_skips
+    run_generator
+    assert_file "test/dummy/config/application.rb", /\s+require\s+["']rails\/all["']/
+    assert_file "test/dummy/config/environments/development.rb" do |content|
+      assert_match(/config\.action_mailer\.raise_delivery_errors = false/, content)
+    end
+    assert_file "test/dummy/config/environments/test.rb" do |content|
+      assert_match(/config\.action_mailer\.delivery_method = :test/, content)
+    end
+    assert_file "test/dummy/config/environments/production.rb" do |content|
+      assert_match(/# config\.action_mailer\.raise_delivery_errors = false/, content)
+    end
+  end
+
   def test_active_record_is_removed_from_frameworks_if_skip_active_record_is_given
     run_generator [destination_root, "--skip-active-record"]
     assert_file "test/dummy/config/application.rb", /#\s+require\s+["']active_record\/railtie["']/
@@ -150,6 +164,20 @@ class PluginGeneratorTest < Rails::Generators::TestCase
     assert_no_file "test/dummy/config/database.yml"
     assert_file "test/test_helper.rb" do |contents|
       assert_no_match(/ActiveRecord/, contents)
+    end
+  end
+
+  def test_action_mailer_is_removed_from_frameworks_if_skip_action_mailer_is_given
+    run_generator [destination_root, "--skip-action-mailer"]
+    assert_file "test/dummy/config/application.rb", /#\s+require\s+["']action_mailer\/railtie["']/
+    assert_file "test/dummy/config/environments/development.rb" do |content|
+      assert_no_match(/config\.action_mailer/, content)
+    end
+    assert_file "test/dummy/config/environments/test.rb" do |content|
+      assert_no_match(/config\.action_mailer/, content)
+    end
+    assert_file "test/dummy/config/environments/production.rb" do |content|
+      assert_no_match(/config\.action_mailer/, content)
     end
   end